requirements.txt
point-cloud-utils==0.30.0
trimesh==4.0.5
numpy-stl==3.1.1
